+++ CHANGELOG.md
@@ -48,6 +48,8 @@ All changes and updates, that are relevant for a user will be documented here
 - show open positions in security accounts overview pages
 - Make it possible to open (create) option transactions
 - Show option positions in account overview
+- Add possibility to make screenshots via shortcuts and menu
+- Add possibility to customize shortcuts
 
 <!-- insertion marker -->
 ## [0.2.3](https://github.com/repo/owner/releases/tag/0.2.3) - 2026-05-17